requirements.txt
fastapi==0.115.6
python-multipart==0.0.19
uvicorn==0.32.1
google-generativeai==0.8.3
python-dotenv==0.21.0